implemented edge detection overlay
The second step of my software, where the edge detection is ran on the base map, and overlaid to work on. It also creates a mask with the new edge overlay. The tools to draw on it are not made yet, but the edge works and has been tested with several different inputs.