made basic mesh exporting
The heightmap now converts into a mesh. It is exported as an OBJ for now, for testing. It has required a lot of iteration with the heightmap page to turn into something usable. It might also require changing things to do more than just arrays of floats. It might require objects that have a float value and other data for processing.