Edge Editing Page added
Now the edge editing page works here as well. It uses text now, so it works much faster than before. It also has added checks to make sure that things don't get advanced improperly(easier for this page, as once the processing is done, we technically have what we need. but there is a warning to make sure users are okay with using the default edges)