added navigation behavior to first page
Now the first page saves a state before going to the next page, so when you go back on the edge page, you can continue where you left off. and you can go back to the welcome page. Warnings are given to the user that data will not be saved on the page you are leaving from.